First pass at creating ics files at build time. #51
