+++ .gitignore
@@ -38,3 +38,11 @@ jspm_packages
 .next
 
 .DS_Store
+
+# Build Pages
+/pages
+/page/*
+
+# ICS Files Generated at Build Time
+/public/download
+/public/download/*
